Renamed BaseCreator to Creator and Creator to StandardCreator.
