+++ tests/creator/manual/_utils/creator/classiccreator.js
@@ -5,7 +5,7 @@
 
 'use strict';
 
-import Creator from '/ckeditor5/creator/creator.js';
+import StandardCreator from '/ckeditor5/creator/standardcreator.js';
 
 import HtmlDataProcessor from '/ckeditor5/engine/dataprocessor/htmldataprocessor.js';
 import Editable from '/ckeditor5/editable.js';
@@ -22,7 +22,7 @@ import ToolbarView from '/ckeditor5/ui/toolbar/toolbarview.js';
 
 import { imitateFeatures, imitateDestroyFeatures } from '../imitatefeatures.js';
 
-export default class ClassicCreator extends Creator {
+export default class ClassicCreator extends StandardCreator {
 	constructor( editor ) {
 		super( editor, new HtmlDataProcessor() );
